Move Cesium from public/cesium to third_party/cesium.
It no longer needs to be mapped publicly.
