+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "public/cesium"]
-	path = public/cesium
+	path = third_party/cesium
 	url = https://github.com/AnalyticalGraphicsInc/cesium.git
