+++ gulpfile.js
@@ -25,11 +25,11 @@ var appJSName = 'ausglobe.js';
 var specJSName = 'ausglobe-specs.js';
 var appEntryJSName = './src/viewer/main.js';
 var workerGlob = [
-    './public/cesium/Source/Workers/*.js',
-    '!./public/cesium/Source/Workers/*.profile.js',
-    '!./public/cesium/Source/Workers/cesiumWorkerBootstrapper.js',
-    '!./public/cesium/Source/Workers/transferTypedArrayTest.js',
-    '!./public/cesium/Source/Workers/createTaskProcessorWorker.js'
+    './third_party/cesium/Source/Workers/*.js',
+    '!./third_party/cesium/Source/Workers/*.profile.js',
+    '!./third_party/cesium/Source/Workers/cesiumWorkerBootstrapper.js',
+    '!./third_party/cesium/Source/Workers/transferTypedArrayTest.js',
+    '!./third_party/cesium/Source/Workers/createTaskProcessorWorker.js'
 ];
 var specGlob = './spec/*.js';
 
@@ -85,7 +85,7 @@ gulp.task('prepare-cesium', ['build-cesium', 'copy-cesium-assets', 'copy-cesiumW
 
 gulp.task('build-cesium', function(cb) {
     return exec('"Tools/apache-ant-1.8.2/bin/ant" build', {
-        cwd : 'public/cesium'
+        cwd : 'third_party/cesium'
     }, function(err, stdout, stderr) {
         if (stderr) {
             console.log('Error while building Cesium: ');
@@ -97,11 +97,11 @@ gulp.task('build-cesium', function(cb) {
 
 gulp.task('copy-cesium-assets', function() {
     return gulp.src([
-            'public/cesium/Source/Workers/transferTypedArrayTest.js',
-            'public/cesium/Source/Assets/**',
-            'public/cesium/Source/Widgets/**/*.css',
-            'public/cesium/Source/Widgets/Images/**'
-        ], { base: 'public/cesium/Source' })
+            'third_party/cesium/Source/Workers/transferTypedArrayTest.js',
+            'third_party/cesium/Source/Assets/**',
+            'third_party/cesium/Source/Widgets/**/*.css',
+            'third_party/cesium/Source/Widgets/Images/**'
+        ], { base: 'third_party/cesium/Source' })
         .pipe(gulp.dest('public/build/Cesium/'));
 });
