+++ src/TableDataSource.js
@@ -11,12 +11,12 @@ And writes a czml file for it to display
 
 //TODO: DOCUMENT using model in GeoJsonDataSource
 
-var defaultValue = require('../public/cesium/Source/Core/defaultValue');
-var CzmlDataSource = require('../public/cesium/Source/DataSources/CzmlDataSource');
-var Color = require('../public/cesium/Source/Core/Color');
-var defineProperties = require('../public/cesium/Source/Core/defineProperties');
-var destroyObject = require('../public/cesium/Source/Core/destroyObject');
-var JulianDate = require('../public/cesium/Source/Core/JulianDate');
+var defaultValue = require('../third_party/cesium/Source/Core/defaultValue');
+var CzmlDataSource = require('../third_party/cesium/Source/DataSources/CzmlDataSource');
+var Color = require('../third_party/cesium/Source/Core/Color');
+var defineProperties = require('../third_party/cesium/Source/Core/defineProperties');
+var destroyObject = require('../third_party/cesium/Source/Core/destroyObject');
+var JulianDate = require('../third_party/cesium/Source/Core/JulianDate');
 
 var EMPTY_OBJECT = {};
