+++ src/viewer/AusGlobeViewer.js
@@ -6,57 +6,57 @@
 "use strict";
 
 /*global require,L,URI,$,XMLDocument,html2canvas,alert,console*/
-var BingMapsApi = require('../../public/cesium/Source/Core/BingMapsApi');
-var BingMapsImageryProvider = require('../../public/cesium/Source/Scene/BingMapsImageryProvider');
-var BingMapsStyle = require('../../public/cesium/Source/Scene/BingMapsStyle');
-var CameraFlightPath = require('../../public/cesium/Source/Scene/CameraFlightPath');
-var Cartesian2 = require('../../public/cesium/Source/Core/Cartesian2');
-var Cartesian3 = require('../../public/cesium/Source/Core/Cartesian3');
-var Cartographic = require('../../public/cesium/Source/Core/Cartographic');
-var CesiumMath = require('../../public/cesium/Source/Core/Math');
-var CesiumTerrainProvider = require('../../public/cesium/Source/Core/CesiumTerrainProvider');
-var Clock = require('../../public/cesium/Source/Core/Clock');
-var ClockRange = require('../../public/cesium/Source/Core/ClockRange');
-var Color = require('../../public/cesium/Source/Core/Color');
-var combine = require('../../public/cesium/Source/Core/combine');
-var Credit = require('../../public/cesium/Source/Core/Credit');
-var DataSourceCollection = require('../../public/cesium/Source/DataSources/DataSourceCollection');
-var defaultValue = require('../../public/cesium/Source/Core/defaultValue');
-var defined = require('../../public/cesium/Source/Core/defined');
-var Entity = require('../../public/cesium/Source/DataSources/Entity');
-var Ellipsoid = require('../../public/cesium/Source/Core/Ellipsoid');
-var EllipsoidTerrainProvider = require('../../public/cesium/Source/Core/EllipsoidTerrainProvider');
-var FeatureDetection = require('../../public/cesium/Source/Core/FeatureDetection');
-var GeographicProjection = require('../../public/cesium/Source/Core/GeographicProjection');
-var CesiumEvent = require('../../public/cesium/Source/Core/Event');
-var Rectangle = require('../../public/cesium/Source/Core/Rectangle');
-var Fullscreen = require('../../public/cesium/Source/Core/Fullscreen');
-var InfoBox = require('../../public/cesium/Source/Widgets/InfoBox/InfoBox');
-var JulianDate = require('../../public/cesium/Source/Core/JulianDate');
-var KeyboardEventModifier = require('../../public/cesium/Source/Core/KeyboardEventModifier');
-var loadJson = require('../../public/cesium/Source/Core/loadJson');
-var loadXML = require('../../public/cesium/Source/Core/loadXML');
-var Material = require('../../public/cesium/Source/Scene/Material');
-var Matrix3 = require('../../public/cesium/Source/Core/Matrix3');
-var Matrix4 = require('../../public/cesium/Source/Core/Matrix4');
-var PolylineCollection = require('../../public/cesium/Source/Scene/PolylineCollection');
-var Rectangle = require('../../public/cesium/Source/Core/Rectangle');
-var RectanglePrimitive = require('../../public/cesium/Source/Scene/RectanglePrimitive');
-var sampleTerrain = require('../../public/cesium/Source/Core/sampleTerrain');
-var SceneMode = require('../../public/cesium/Source/Scene/SceneMode');
-var ScreenSpaceEventHandler = require('../../public/cesium/Source/Core/ScreenSpaceEventHandler');
-var ScreenSpaceEventType = require('../../public/cesium/Source/Core/ScreenSpaceEventType');
-var Transforms = require('../../public/cesium/Source/Core/Transforms');
-var Tween = require('../../public/cesium/Source/ThirdParty/Tween');
-var Viewer = require('../../public/cesium/Source/Widgets/Viewer/Viewer');
-var viewerEntityMixin = require('../../public/cesium/Source/Widgets/Viewer/viewerEntityMixin');
-var WebMapServiceImageryProvider = require('../../public/cesium/Source/Scene/WebMapServiceImageryProvider');
-var WebMercatorProjection = require('../../public/cesium/Source/Core/WebMercatorProjection');
-var when = require('../../public/cesium/Source/ThirdParty/when');
+var BingMapsApi = require('../../third_party/cesium/Source/Core/BingMapsApi');
+var BingMapsImageryProvider = require('../../third_party/cesium/Source/Scene/BingMapsImageryProvider');
+var BingMapsStyle = require('../../third_party/cesium/Source/Scene/BingMapsStyle');
+var CameraFlightPath = require('../../third_party/cesium/Source/Scene/CameraFlightPath');
+var Cartesian2 = require('../../third_party/cesium/Source/Core/Cartesian2');
+var Cartesian3 = require('../../third_party/cesium/Source/Core/Cartesian3');
+var Cartographic = require('../../third_party/cesium/Source/Core/Cartographic');
+var CesiumMath = require('../../third_party/cesium/Source/Core/Math');
+var CesiumTerrainProvider = require('../../third_party/cesium/Source/Core/CesiumTerrainProvider');
+var Clock = require('../../third_party/cesium/Source/Core/Clock');
+var ClockRange = require('../../third_party/cesium/Source/Core/ClockRange');
+var Color = require('../../third_party/cesium/Source/Core/Color');
+var combine = require('../../third_party/cesium/Source/Core/combine');
+var Credit = require('../../third_party/cesium/Source/Core/Credit');
+var DataSourceCollection = require('../../third_party/cesium/Source/DataSources/DataSourceCollection');
+var defaultValue = require('../../third_party/cesium/Source/Core/defaultValue');
+var defined = require('../../third_party/cesium/Source/Core/defined');
+var Entity = require('../../third_party/cesium/Source/DataSources/Entity');
+var Ellipsoid = require('../../third_party/cesium/Source/Core/Ellipsoid');
+var EllipsoidTerrainProvider = require('../../third_party/cesium/Source/Core/EllipsoidTerrainProvider');
+var FeatureDetection = require('../../third_party/cesium/Source/Core/FeatureDetection');
+var GeographicProjection = require('../../third_party/cesium/Source/Core/GeographicProjection');
+var CesiumEvent = require('../../third_party/cesium/Source/Core/Event');
+var Rectangle = require('../../third_party/cesium/Source/Core/Rectangle');
+var Fullscreen = require('../../third_party/cesium/Source/Core/Fullscreen');
+var InfoBox = require('../../third_party/cesium/Source/Widgets/InfoBox/InfoBox');
+var JulianDate = require('../../third_party/cesium/Source/Core/JulianDate');
+var KeyboardEventModifier = require('../../third_party/cesium/Source/Core/KeyboardEventModifier');
+var loadJson = require('../../third_party/cesium/Source/Core/loadJson');
+var loadXML = require('../../third_party/cesium/Source/Core/loadXML');
+var Material = require('../../third_party/cesium/Source/Scene/Material');
+var Matrix3 = require('../../third_party/cesium/Source/Core/Matrix3');
+var Matrix4 = require('../../third_party/cesium/Source/Core/Matrix4');
+var PolylineCollection = require('../../third_party/cesium/Source/Scene/PolylineCollection');
+var Rectangle = require('../../third_party/cesium/Source/Core/Rectangle');
+var RectanglePrimitive = require('../../third_party/cesium/Source/Scene/RectanglePrimitive');
+var sampleTerrain = require('../../third_party/cesium/Source/Core/sampleTerrain');
+var SceneMode = require('../../third_party/cesium/Source/Scene/SceneMode');
+var ScreenSpaceEventHandler = require('../../third_party/cesium/Source/Core/ScreenSpaceEventHandler');
+var ScreenSpaceEventType = require('../../third_party/cesium/Source/Core/ScreenSpaceEventType');
+var Transforms = require('../../third_party/cesium/Source/Core/Transforms');
+var Tween = require('../../third_party/cesium/Source/ThirdParty/Tween');
+var Viewer = require('../../third_party/cesium/Source/Widgets/Viewer/Viewer');
+var viewerEntityMixin = require('../../third_party/cesium/Source/Widgets/Viewer/viewerEntityMixin');
+var WebMapServiceImageryProvider = require('../../third_party/cesium/Source/Scene/WebMapServiceImageryProvider');
+var WebMercatorProjection = require('../../third_party/cesium/Source/Core/WebMercatorProjection');
+var when = require('../../third_party/cesium/Source/ThirdParty/when');
 
-var knockout = require('../../public/cesium/Source/ThirdParty/knockout');
+var knockout = require('../../third_party/cesium/Source/ThirdParty/knockout');
 var komapping = require('../../public/third_party/knockout.mapping');
-var knockoutES5 = require('../../public/cesium/Source/ThirdParty/knockout-es5');
+var knockoutES5 = require('../../third_party/cesium/Source/ThirdParty/knockout-es5');
 
 var corsProxy = require('../corsProxy');
 var GeoDataBrowser = require('./GeoDataBrowser');
