+++ src/viewer/GeoDataBrowser.js
@@ -2,11 +2,11 @@
 
 /*global require*/
 
-var defineProperties = require('../../public/cesium/Source/Core/defineProperties');
-var getElement = require('../../public/cesium/Source/Widgets/getElement');
+var defineProperties = require('../../third_party/cesium/Source/Core/defineProperties');
+var getElement = require('../../third_party/cesium/Source/Widgets/getElement');
 
-var knockout = require('../../public/cesium/Source/ThirdParty/knockout');
-var knockoutES5 = require('../../public/cesium/Source/ThirdParty/knockout-es5');
+var knockout = require('../../third_party/cesium/Source/ThirdParty/knockout');
+var knockoutES5 = require('../../third_party/cesium/Source/ThirdParty/knockout-es5');
 
 var GeoDataBrowserViewModel = require('./GeoDataBrowserViewModel');
