+++ src/workerLoader.js
@@ -1,7 +1,7 @@
 /*global require*/
 "use strict";
 
-var createTaskProcessorWorker = require('../public/cesium/Source/Workers/createTaskProcessorWorker');
+var createTaskProcessorWorker = require('../third_party/cesium/Source/Workers/createTaskProcessorWorker');
 
 module.exports = function(workerName) {
     return createTaskProcessorWorker.workers[workerName];
